+++ test/core/chain.test.js
@@ -21,4 +21,24 @@ suite('chain', function() {
     assert.ok(!parser.parse('x').status);
   });
 
+  test('should pass state through', function() {
+    function test(n) {
+      return Parsimmon(function(input, i, state) {
+        assert.strictEqual(state, n);
+        return Parsimmon.makeSuccess(i, undefined, state);
+      });
+    }
+    function always10() {
+      return 10;
+    }
+    var inc = Parsimmon(function(input, i, state) {
+      return Parsimmon.makeSuccess(i, null, state + 1);
+    });
+    Parsimmon.any.chain(function() {
+      return test()
+    })
+    test(0).then(Parsimmon.any.map(always10)).skip(test(0)).tryParse('a', 0);
+    inc.then(Parsimmon.any.map(always10)).skip(test(1)).tryParse('b', 0);
+  });
+
 });