+++ test/core/makeFailure.test.js
@@ -3,12 +3,13 @@
 test('Parsimmon.makeFailure', function() {
   var furthest = 4444;
   var expected = 'waiting in the clock tower';
-  var result = Parsimmon.makeFailure(furthest, expected);
+  var result = Parsimmon.makeFailure(furthest, expected, undefined);
   assert.deepEqual(result, {
     status: false,
     index: -1,
     value: null,
     furthest: furthest,
-    expected: [expected]
+    expected: [expected],
+    state: undefined
   });
 });