+++ test/core/makeSuccess.test.js
@@ -3,12 +3,13 @@
 test('Parsimmon.makeSuccess', function() {
   var index = 11;
   var value = 'a lucky number';
-  var result = Parsimmon.makeSuccess(index, value);
+  var result = Parsimmon.makeSuccess(index, value, 'state');
   assert.deepEqual(result, {
     status: true,
     index: index,
     value: value,
     furthest: -1,
-    expected: []
+    expected: [],
+    state: 'state'
   });
 });