+++ test/core/map.test.js
@@ -23,4 +23,21 @@ suite('map', function() {
     });
   });
 
+  test('should pass state through', function() {
+    function test(n) {
+      return Parsimmon(function(input, i, state) {
+        assert.strictEqual(state, n);
+        return Parsimmon.makeSuccess(i, undefined, state);
+      });
+    }
+    function always10() {
+      return 10;
+    }
+    var inc = Parsimmon(function(input, i, state) {
+      return Parsimmon.makeSuccess(i, null, state + 1);
+    });
+    test(0).then(Parsimmon.any.map(always10)).skip(test(0)).tryParse('a', 0);
+    inc.then(Parsimmon.any.map(always10)).skip(test(1)).tryParse('b', 0);
+  });
+
 });