+++ test/core/notFollowedBy.test.js
@@ -33,4 +33,23 @@ suite('Parsimmon.notFollowedBy', function() {
     assert.deepEqual(answer.value, ['abc', 'd']);
   });
 
+  test('should pass state through unchanged', function() {
+    function test(n) {
+      return Parsimmon(function(input, i, state) {
+        assert.strictEqual(state, n);
+        return Parsimmon.makeSuccess(i, undefined, state);
+      });
+    }
+    var inc = Parsimmon(function(input, i, state) {
+      return Parsimmon.makeSuccess(i, null, state + 1);
+    });
+    var parser = Parsimmon.notFollowedBy(Parsimmon.regexp(/a/));
+    test(0)
+      .then(inc)
+      .then(parser)
+      .skip(test(1))
+      .then(Parsimmon.string('b'))
+      .tryParse('b', 0);
+  });
+
 });